requirements.txt
tensorflow-cpu
opencv-python
numpy
pandas
matplotlib
scikit-learn
flask
Pillow
h5py
seaborn
tqdm
